requirements.txt
selenium
numpy
pandas
scikit-learn
matplotlib
seaborn
fastapi
pymongo
python-dotenv